requirements.txt
colorama
scapy
icmplib
platform
python-whois
ip2geotools
